updates according to comments
1 add tests for ConsumeTriggers and ConsumeAppMetric in aggregator
2 metric-poller share only one httpclient
3 add test for metric-server is not running in MetricPoller
4 other small updates

+++ src/dataaggregator/appmetric/appmetric.go
@@ -3,7 +3,7 @@ package appmetric
 type AppMonitor struct {
 	AppId          string
 	MetricType     string
-	StatWindowSecs int64
+	StatWindowSecs int
 }
 type AppMetric struct {
 	AppId      string
